Codebase V2025.2
Updated Single Motor Subsytem and added debugging info array to 'Motors.java'. This was to allow us to view any debugging information as needed. It needs to be tested in order to see how it displays. They were put as strings so they could all be displayed between multiple different types.